requirements.txt
apitally==0.11.2
asgiref==3.8.1
backoff==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==3.14.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-jazzmin==2.6.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
packaging==24.0
Pillow==9.5.0
prometheus_client==0.20.0
PyJWT==2.6.0
PyPDF2==3.0.1
pytz==2023.3
PyYAML==6.0.1
redis==5.0.8
requests==2.31.0
setuptools==72.2.0
sqlparse==0.4.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
